@@ -1,4 +1,4 @@
-lib = File.expand_path('../lib/', __FILE__)
+lib = File.expand_path('lib', __dir__)
 $LOAD_PATH.unshift lib unless $LOAD_PATH.include?(lib)
 
 require 'spree_reviews/version'
@@ -9,38 +9,26 @@ Gem::Specification.new do |s|
   s.version     = SpreeReviews.version
   s.summary     = 'Basic review and ratings facility for Spree'
   s.description = s.summary
-  s.required_ruby_version = '>= 2.1.0'
+  s.required_ruby_version = '>= 2.5.0'
 
-  s.authors 	   = ['Paul Callaghan']
-  s.email        = 'paulcc.two@gmail.com'
-  s.homepage     = 'https://github.com/spree-contrib/spree-reviews/'
-  s.license      = 'BSD-3'
+  s.author    = ['Paul Callaghan']
+  s.email     = 'paulcc.two@gmail.com'
+  s.homepage  = 'https://github.com/spree-contrib/spree-reviews/'
+  s.license = 'BSD-3'
 
-  s.files        = `git ls-files`.split("\n")
-  s.test_files   = `git ls-files -- spec/*`.split("\n")
+  s.files = `git ls-files`.split("\n").reject { |f| f.match(/^spec/) && !f.match(/^spec\/fixtures/) }
   s.require_path = 'lib'
   s.requirements << 'none'
 
-  spree_version = '>= 3.1.0', '< 4.0'
-  s.add_runtime_dependency 'spree_core', spree_version
-  s.add_runtime_dependency 'spree_auth_devise', spree_version
-  s.add_runtime_dependency 'spree_extension'
+  spree_version = '>= 4.5'
+  s.add_dependency 'deface', '~> 1.0'
+  s.add_dependency 'spree_api', spree_version
+  s.add_dependency 'spree_backend', spree_version
+  s.add_dependency 'spree_frontend', spree_version
+  s.add_dependency 'spree_core', spree_version
+  s.add_dependency 'spree_extension'
 
-  s.add_development_dependency 'ffaker'
-  s.add_development_dependency 'capybara'
-  s.add_development_dependency 'capybara-screenshot'
-  s.add_development_dependency 'database_cleaner'
-  s.add_development_dependency 'poltergeist'
-  s.add_development_dependency 'rspec-rails'
-  s.add_development_dependency 'factory_girl'
-  s.add_development_dependency 'sqlite3'
-  s.add_development_dependency 'simplecov'
-  s.add_development_dependency 'coffee-rails'
   s.add_development_dependency 'sass-rails'
-  s.add_development_dependency 'guard-rspec'
-  s.add_development_dependency 'pry-rails'
-  s.add_development_dependency 'rubocop'
-  s.add_development_dependency 'appraisal'
+  s.add_development_dependency 'spree_dev_tools'
   s.add_development_dependency 'pg'
-  s.add_development_dependency 'mysql2'
 end
