@@ -7,7 +7,7 @@ class Spree::FeedbackReview < ActiveRecord::Base
     only_integer: true,
     greater_than_or_equal_to: 1,
     less_than_or_equal_to: 5,
-    message: Spree.t(:you_must_enter_value_for_rating)
+    message: :you_must_enter_value_for_rating
   }
 
   default_scope { most_recent_first }
