@@ -1,18 +1,24 @@
 # Add access to reviews/ratings to the product model
-Spree::Product.class_eval do
-  has_many :reviews
-
-  def stars
-    avg_rating.try(:round) || 0
-  end
-
-  def recalculate_rating
-    self[:reviews_count] = reviews.reload.approved.count
-    if reviews_count > 0
-      self[:avg_rating] = reviews.approved.sum(:rating).to_f / reviews_count
-    else
-      self[:avg_rating] = 0
+module Spree
+  module ProductDecorator
+    def self.prepended(base)
+      base.has_many :reviews
     end
-    save
+
+    def stars
+      avg_rating.try(:round) || 0
+    end
+
+    def recalculate_rating
+      self[:reviews_count] = reviews.reload.approved.count
+      self[:avg_rating] = if reviews_count > 0
+                            reviews.approved.sum(:rating).to_f / reviews_count
+                          else
+                            0
+                          end
+      save
+    end
+
+    ::Spree::Product.prepend self if ::Spree::Product.included_modules.exclude?(self)
   end
 end
