@@ -11,7 +11,7 @@ class Spree::Review < ActiveRecord::Base
     only_integer: true,
     greater_than_or_equal_to: 1,
     less_than_or_equal_to: 5,
-    message: Spree.t(:you_must_enter_value_for_rating)
+    message: :you_must_enter_value_for_rating
   }
 
   default_scope { order('spree_reviews.created_at DESC') }
@@ -26,6 +26,7 @@ class Spree::Review < ActiveRecord::Base
 
   def feedback_stars
     return 0 if feedback_reviews.size <= 0
+
     ((feedback_reviews.sum(:rating) / feedback_reviews.size) + 0.5).floor
   end
 
