@@ -1,12 +1,20 @@
-RSpec.feature 'Admin Settings for Reviews', :js do
+require 'spec_helper'
+
+describe 'Admin Settings for Reviews', :js do
   stub_authorization!
 
-  background do
+  before do
     visit spree.edit_admin_review_settings_path
   end
 
-  scenario 'update' do
-    expect(page).to have_text_like 'Size of the review snippets'
+  it 'update' do
+    the_key_string = Spree.t(:preview_size, scope: :spree_reviews)
+
+    if Spree.version.to_f < 4.0
+      expect(page).to have_content(the_key_string.upcase)
+    else
+      expect(page).to have_content(the_key_string)
+    end
 
     check 'include_unapproved_reviews'
     check 'feedback_rating'
@@ -15,9 +23,10 @@ RSpec.feature 'Admin Settings for Reviews', :js do
     check 'track_locale'
     check 'show_identifier'
     fill_in 'preview_size', with: '5'
+    fill_in 'paginate_size', with: '6'
     click_button 'Update'
 
-    expect(page).to have_text 'successfully updated!'
+    expect(page).to have_content('successfully updated!')
 
     setting = Spree::ReviewSetting.new
 
