@@ -1,32 +1,10 @@
-require 'simplecov'
-SimpleCov.start 'rails'
+# Configure Rails Environment
+ENV['RAILS_ENV'] = 'test'
 
-ENV['RAILS_ENV'] ||= 'test'
+require File.expand_path('dummy/config/environment.rb', __dir__)
 
-begin
-  require File.expand_path('../dummy/config/environment', __FILE__)
-rescue LoadError
-  puts 'Could not load dummy application. Please ensure you have run `bundle exec rake test_app`'
-  exit
-end
+require 'spree_dev_tools/rspec/spec_helper'
 
-require 'rspec/rails'
-require 'ffaker'
-require 'pry'
-
-RSpec.configure do |config|
-  config.fail_fast = false
-  config.filter_run focus: true
-  config.run_all_when_everything_filtered = true
-
-  config.mock_with :rspec
-  config.use_transactional_fixtures = false
-  config.raise_errors_for_deprecations!
-  config.infer_spec_type_from_file_location!
-
-  config.expect_with :rspec do |expectations|
-    expectations.syntax = :expect
-  end
-end
-
-Dir[File.join(File.dirname(__FILE__), 'support/**/*.rb')].each { |file| require file }
+# Requires supporting ruby files with custom matchers and macros, etc,
+# in spec/support/ and its subdirectories.
+Dir[File.join(File.dirname(__FILE__), 'support/**/*.rb')].sort.each { |f| require f }
